homework 1 was fixed

+++ project/include/utils.h
@@ -3,7 +3,7 @@
 #include <stdio.h>
 
 size_t timer_from(int from);
-int custom_pow(int base, int power);
+double custom_pow(int base, int power);
 
 #endif  // PROJECT_INCLUDE_UTILS_H_
